+++ src/test/java/org/folio/rest/api/PatronNoticePoliciesApiTest.java
@@ -54,9 +54,35 @@ public class PatronNoticePoliciesApiTest extends ApiTests {
       .put("realTime", "true")
       .put("sendOptions", sendOptions);
 
+    JsonObject manualDueDateChangeNotice = new JsonObject()
+      .put("templateId", UUID.randomUUID().toString())
+      .put("format", "Email")
+      .put("realTime", "true")
+      .put("sendOptions", new JsonObject()
+        .put("sendWhen", "Manual due date change"));
+
+    JsonObject itemRecalledNotice = new JsonObject()
+      .put("templateId", UUID.randomUUID().toString())
+      .put("format", "Email")
+      .put("realTime", "true")
+      .put("sendOptions", new JsonObject()
+        .put("sendWhen", "Item recalled"));
+
+    JsonObject holdExpirationChangeNotice = new JsonObject()
+      .put("templateId", UUID.randomUUID().toString())
+      .put("format", "Email")
+      .put("realTime", "true")
+      .put("sendOptions", new JsonObject()
+        .put("sendWhen", "Hold expiration"));
+
     JsonObject noticePolicy = new JsonObject()
       .put("name", "sample policy")
-      .put("requestNotices", new JsonArray().add(requestNotice));
+      .put("requestNotices", new JsonArray()
+        .add(requestNotice)
+        .add(holdExpirationChangeNotice))
+      .put("loanNotices", new JsonArray()
+        .add(manualDueDateChangeNotice)
+        .add(itemRecalledNotice));
 
     JsonResponse response = postPatronNoticePolicy(noticePolicy);