+++ src/test/java/org/folio/rest/api/StorageTestSuite.java
@@ -53,7 +53,8 @@ import io.vertx.ext.sql.ResultSet;
   PatronNoticePoliciesApiTest.class,
   RequestPoliciesApiTest.class,
   RequestExpirationApiTest.class,
-  ScheduledNoticesAPITest.class
+  ScheduledNoticesAPITest.class,
+  PatronActionSessionAPITest.class
 })
 
 public class StorageTestSuite {